Make sure to focus inner text field when creating / rebuilding edit fields in datetime widget.
The blur is a regression from bug 1740989, but it's really uncovering a
regression from bug 1729342, sorta. Before that, we used to focus the
inner text field on rebuild via focusInnerTextBox().

One could argue that the focus fixup rule should really deal with this,
but I guess since content can't really unattach shadow DOM, it's less of
a general issue and more of an implementation detail.

The test uncovers another pre-existing issue (switching from date ->
test was also broken, and we left a bogus ElementState::FOCUS state in
the element).

Differential Revision: https://phabricator.services.mozilla.com/D158167

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1792057
gecko-commit: 4edc81a94eaa522befc096d01708a7217e3109fa
gecko-reviewers: smaug