Ignore presentation time if visibility changed, for Event Timing
EventTiming registers for presentation feedback times for events which are expected to have a visual update.  However, if page visibility changes, that can affect the time it takes to present the next paint update.

We were already ignoring the time to next paint portion for UKM reporting of interactions, but now we also ignore this time for Event Timing reporting to Web Performance timeline.

Bug: 1312568
Change-Id: Iee0b5eee81fb216a96ecbfcc41ca26601ab6ba1b