[shared storage] add permissions policy for individual output gate: selectURL() and privateAggregation
This allows more fine-grained control of the output gates

PR: https://github.com/WICG/shared-storage/pull/52

This CL only implement the PAA check for shared storage part. For FLEDGE, it will be done in a separate CL.

Also add the new features in
kFencedFrameOpaqueAdsDefaultAllowedFeatures: they are supposed to
be used from within the FF.

The private aggregation permissions policy is tested in a browser test / is not easily testable via web_test: right now, only selectURL()’s result rendered in FF can expose state in the worklet; but if the permissions policy is disallowed in the embedder, the FF navigation will fail (as the permissions is in kFencedFrameOpaqueAdsDefaultAllowedFeatures).

Bug: 1400840
Change-Id: I74a285827126f5cf8a43cb780e37693f4c0c7ddf