[ComputePressure] Resolve/reject the promises correctly when calling observe method.
In current implementation, when we call observe method on the same
observer like this:
    let promise1 = observer.observe('cpu');
    let promise2 = observer.observe('cpu');
or when we call observe method on different observers like this:
    let promise1 = observer1.observe('cpu');
    let promise2 = observer2.observe('cpu');
promise2 is resolved before promise1 no matter promise1 is resolved
or rejected. This CL resolves this issue by making all promises
resolved/rejected when PromiseObserverManager notifies
PressureObserver.

Bug: 1393210, 1360396
Change-Id: I580953c6ff586bd517eae610d695c4e55ff80ded