[Web Payment] Remove identity from "canmakepayment" event
Before this patch, the "canmakepayment" event for service worker payment
handlers contained user and merchant identifiers, which could be removed
by either enrolling in an OT or by enabling the
chrome://flags/#clear-identity-in-can-make-payment flag.

This patch:
1) Makes the identity fields in "canmakepayment" be removed by default.
2) Replaces the chrome://flags/#clear-identity-in-can-make-payment flag
   with the chrome://flags/#add-identity-in-can-make-payment flag.
3) Adds an OT for temporarily re-enabling the identity passing.
4) Updates test expectations, include WPTs.
This patch does not remove the old flag and OT until all of the
dependent code is updated.

After this patch, the "canmakepayment" event for service worker payment
handlers does not contain user or merchant identifiers, unless the
chrome://flags/#add-identity-in-can-make-payment flag is enabled or the
website is enrolled in OT. If the identifiers are accessed (because
either the flag or OT re-enables them), then DevTools continues to print
a deprecation message.

Chrome status:
https://chromestatus.com/feature/5190978431352832

Intent:
https://groups.google.com/a/chromium.org/g/blink-dev/c/AM2bwKxXacQ/m/UPPjcrDvDQAJ

Spec change:
https://github.com/w3c/payment-handler/pull/404/files

Bug: 1290492
Change-Id: I81d89150cf118614c26b4c620a54d8f9a021693f